problem 1 marked

+++ cpp/ps0001/pc0001.cpp
@@ -2,9 +2,13 @@
 using namespace std;
 int main()
 {
-    int num1 , num2 , num3;
+    // promptiung user for 3 numbers to add 1/2
+    short num1 , num2 , num3;
     cout << "Enter Three numbers: ";
     cin >> num1 >> num2 >> num3;
+
+    // outputting the result of the summation process 3/3
     cout << num1 + num2 + num3 << endl;
     return 0;
 }
+// good work 👍 7/10
